Move AdaptiveParagraph and AdaptiveTextRun to their own files.
