+++ source/dotnet/Library/AdaptiveCards/AdaptiveRichTextBlock.cs
@@ -20,7 +20,6 @@ namespace AdaptiveCards
 
         public AdaptiveRichTextBlock()
         {
-
         }
 
         /// <summary>
@@ -59,99 +58,5 @@ namespace AdaptiveCards
         [XmlElement(typeof(AdaptiveParagraph))]
 #endif
         public List<AdaptiveParagraph> Paragraphs { get; set; } = new List<AdaptiveParagraph>();
-
-
-    }
-
-    [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
-    public class AdaptiveParagraph
-    {
-        [JsonRequired]
-        [JsonConverter(typeof(IgnoreEmptyItemsConverter<AdaptiveTextRun>))]
-#if !NETSTANDARD1_3
-        [XmlElement(typeof(AdaptiveTextRun))]
-#endif
-        public List<AdaptiveTextRun> Inlines { get; set; } = new List<AdaptiveTextRun>();
-
-
-    }
-
-#if !NETSTANDARD1_3
-    [XmlType(TypeName = AdaptiveTextBlock.TypeName)]
-#endif
-    public class AdaptiveTextRun : AdaptiveTypedElement, ITextElement
-    {
-        public const string TypeName = "TextRun";
-
-        public override string Type { get; set; } = TypeName;
-
-        public AdaptiveTextRun()
-        {
-
-        }
-
-        public AdaptiveTextRun(string text)
-        {
-            Text = text;
-        }
-
-        /// <summary>
-        ///     The size of the text
-        /// </summary>
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-#if !NETSTANDARD1_3
-        [XmlAttribute]
-#endif
-        [DefaultValue(typeof(AdaptiveTextSize), "normal")]
-        public AdaptiveTextSize Size { get; set; }
-
-        /// <summary>
-        ///     The weight of the text
-        /// </summary>
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-#if !NETSTANDARD1_3
-        [XmlAttribute]
-#endif
-        [DefaultValue(typeof(AdaptiveTextWeight), "normal")]
-        public AdaptiveTextWeight Weight { get; set; }
-
-        /// <summary>
-        ///     The color of the text
-        /// </summary>
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-#if !NETSTANDARD1_3
-        [XmlAttribute]
-#endif
-        [DefaultValue(typeof(AdaptiveTextColor), "default")]
-        public AdaptiveTextColor Color { get; set; }
-
-        /// <summary>
-        ///     Make the text less prominent
-        /// </summary>
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-#if !NETSTANDARD1_3
-        [XmlAttribute]
-#endif
-        [DefaultValue(false)]
-        public bool IsSubtle { get; set; }
-
-        /// <summary>
-        ///     The text to display
-        /// </summary>
-        [JsonRequired]
-#if !NETSTANDARD1_3
-        [XmlText]
-#endif
-        public string Text { get; set; } = " ";
-
-        /// <summary>
-        ///     The font style of the text
-        /// </summary>
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore)]
-#if !NETSTANDARD1_3
-        [XmlAttribute]
-#endif
-        [DefaultValue(typeof(AdaptiveFontStyle), "default")]
-        public AdaptiveFontStyle FontStyle { get; set; }
     }
 }
